clean code, rename fn

+++ src/transaction/generator.rs
@@ -21,7 +21,7 @@ pub fn random() -> Transaction {
 fn tx_input() -> Input {
     let mut rng = rand::thread_rng();
     let hash = crypto::generator::h256();
-    let index = rng.next_u32();
+    let index = rng.next_u32() as usize;
     let value = rng.next_u64();
     let recipient = crypto::generator::h256();
     return Input {
